1) Added a bit of performance to calculate line checksum
Unfortunatelly, after 600 bytes of text, in one line or more lines, the system slows down the uart reading. So I think thats a big problem to the project.

I think using LCD could be faster, but idk...may be, I´m able to create an text editor, but not able to be fast as possible into this little machine.

So Sad :(